requirements.txt
pytest==7.2.2
selenium==4.8.2
allure-pytest==2.13.0
requests==2.28.2
pytest-xdist==3.2.0